Zord-vision/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
opencv-python==4.10.0.82
numpy==1.26.0
mediapipe==0.10.14
google-genai==0.8.0
face-recognition==2.1.7
python-dotenv==1.0.0
